backend/requirements.txt
django-cors-headers
Django==4.2.1
django-colorful==1.3
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.5.0
drf-spectacular
flake8==6.0.0
gunicorn==20.1.0
importlib-metadata==4.12.
isort
Markdown==3.4.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
reportlab==3.6.1
zipp==3.15.0